fix imports and some rearrangement of files & modules
