+++ Cargo.toml
@@ -2,4 +2,4 @@
 [workspace]
 edition = "2021"
 resolver = "2"
-members = ["zeal", "zeal-ast", "zeal-core", "zeal-llvm"]
+members = ["zeal", "zeal-ast", "zeal-core", "zeal-llvm", "zvm"]
