+++ zvm/src/err.rs
@@ -1,22 +1,21 @@
-use std::collections::HashMap;
-
-use zeal_core::{str::ZIdent, val::ZValue};
+use std::{collections::HashMap, rc::Rc};
 
 use crate::{
     opcode::{Op, Opcode, VarOp},
     state::Scope,
+    val::Val,
 };
 
 #[derive(Debug, thiserror::Error, Clone)]
 pub enum CompileError {
     #[error("Unresolved local identifier: {name}. Compiling Op: {op}, locals: {scope_state}")]
     UnresolvedLocal {
-        name: ZIdent,
+        name: Rc<str>,
         op: Op,
         scope_state: Scope,
     },
     #[error("Invalid Assigment on Identifier: {name}, var_op: {op_ty} ")]
-    InvalidAssignment { op_ty: VarOp, name: ZIdent },
+    InvalidAssignment { op_ty: VarOp, name: Rc<str> },
 }
 
 #[derive(Debug, thiserror::Error, Clone)]
@@ -32,9 +31,9 @@ pub enum RuntimeError {
             "Unknown Identifier: {name}, encountered from opcode: {opcode:?}, vm constants: {constants:?}, globals: {globals:?}"
         )]
     VMUnknownIdentifier {
-        name: ZIdent,
+        name: Rc<str>,
         opcode: Opcode,
-        constants: Vec<ZValue>,
-        globals: HashMap<ZIdent, ZValue>,
+        constants: Vec<Val>,
+        globals: HashMap<String, Val>,
     },
 }
